+++ src/main/java/com/mahait/gov/in/service/RegularReportService.java
@@ -14,7 +14,7 @@ public interface RegularReportService {
 
 	List<Object[]> findbillgrp(Long billno);
 
-	List<Object[]> findpaybill(int billNumber, int monthName, int yearName, String ddo);
+	List<Object[]> findpaybill(Long billNumber, int monthName, int yearName, String ddo);
 
 	String getbillGroup(int billnumber);
